refactor: improving Tabs/Tab components
