+++ src/components/Tabs/Tabs.js
@@ -1,9 +1,17 @@
+import { useState } from 'react';
 import TabsContext from '../TabsContext';
 import styles from './Tabs.module.css';
 
 const Tabs = ({ value, children }) => {
+  const [activeTab, setActiveTab] = useState(value);
+
+  const contextValue = {
+    currentValue: activeTab,
+    setActiveTab,
+  };
+
   return (
-    <TabsContext.Provider value={value}>
+    <TabsContext.Provider value={contextValue}>
       <nav className={styles.TabsContainer}>
         {children.map(tab => {
           return (
